Consolidate signUp tests
